chat ui & logic supplement

+++ frontend/src/app/components/home/home.component.ts
@@ -154,7 +154,9 @@ export class HomeComponent {
   }
 
   logout() {
-    this.authService.removeToken()
-    this.router.navigate(['sign-in'])
+    if(confirm("Logout ?")){
+      this.authService.removeToken()
+      this.router.navigate(['sign-in'])
+    }
   }
 }
